requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-multipart
pydantic
alembic
python-dotenv
pydantic[email]
passlib
bcrypt==4.3.0